Consider places where this might be important.
Sometimes trailing whitespace is required. In these cases, my
configuration shouldn't be removing that data. However, it should
also be obvious to a developer using the file that the whitespace
is important - or it's just as realistic that they might manually
remove unnecessary whitespace anyway.

Solution? Stop putting whitespace at the end of your files without
denoting it specifically with a backslash.